+++ package.json
@@ -162,6 +162,7 @@
     "firebase-tools": "^13.13.2",
     "husky": "^9.0.11",
     "lint-staged": "^15.2.2",
+    "madge": "^8.0.0",
     "mocha": "^10.3.0",
     "npm-run-all": "^4.1.5",
     "secretlint": "^8.2.4",
