+++ src/functionSchema/functionSchemaTypes.ts
@@ -1,0 +1,3 @@
+// In this file to break circular dependencies
+/** The name of the func decorator. */
+export const FUNC_DECORATOR_NAME = 'func';
