+++ src/modules/firestore/firestoreApplicationContext.ts
@@ -4,7 +4,7 @@ import { FirestoreCodeReviewService } from '#firestore/firestoreCodeReviewServic
 import { FirestoreCacheService } from '#firestore/firestoreFunctionCacheService';
 import { FirestoreLlmCallService } from '#firestore/firestoreLlmCallService';
 import { FirestoreUserService } from '#firestore/firestoreUserService';
-import { ApplicationContext } from '../../applicationContext';
+import { ApplicationContext } from '../../applicationTypes';
 
 export function firestoreApplicationContext(): ApplicationContext {
 	return {
