+++ src/modules/memory/inMemoryApplicationContext.ts
@@ -4,7 +4,7 @@ import { InMemoryCodeReviewService } from '#modules/memory/inMemoryCodeReviewSer
 import { InMemoryFunctionCacheService } from '#modules/memory/inMemoryFunctionCacheService';
 import { InMemoryLlmCallService } from '#modules/memory/inMemoryLlmCallService';
 import { InMemoryUserService } from '#modules/memory/inMemoryUserService';
-import { ApplicationContext } from '../../applicationContext';
+import { ApplicationContext } from '../../applicationTypes';
 
 export function inMemoryApplicationContext(): ApplicationContext {
 	return {
