+++ src/swe/prompt.ts
@@ -4,3 +4,11 @@ Before working on a task think carefully and detail what assumptions you can mak
 
 Here is the information you need for the current step of the task:
 `;
+
+export function buildPrompt(args: {
+	information: string;
+	requirements: string;
+	action: string;
+}): string {
+	return `${basePrompt}\n\n${args.information}\n\nThe requirements of the task are as follows:\n<requirements>\n${args.requirements}\n</requirements>\n\nThe action to be performed is as follows:\n<action>\n${args.action}\n</action>\n`;
+}
