+++ variables/local.env.example
@@ -28,18 +28,20 @@ HIL_BUDGET=1
 HIL_COUNT=5
 
 ANTHROPIC_API_KEY=
-OPENAI_API_KEY=
-GROQ_API_KEY=
-TOGETHERAI_API_KEY=
-FIREWORKS_API_KEY=
+CEREBRAS_API_KEY=
 DEEPSEEK_API_KEY=
 DEEPINFRA_API_KEY=
-CEREBRAS_API_KEY=
-XAI_API_KEY=
-DEEPINFRA_API_KEY=
+FIREWORKS_API_KEY=
+GEMINI_API_KEY=
+GROQ_API_KEY=
 NEBIUS_API_KEY=
-SAMBANOVA_API_KEY=
 #OLLAMA_API_URL=http://localhost:11434
+OPENAI_API_KEY=
+OPENROUTER_API_KEY=
+SAMBANOVA_API_KEY=
+TOGETHERAI_API_KEY=
+XAI_API_KEY=
+
 
 GITLAB_TOKEN=
 GITLAB_HOST=www.gitlab.com # OR your self-hosted domain
