+++ variables/test.env
@@ -19,6 +19,7 @@ GCLOUD_REGION=us-central1
 GCLOUD_CLAUDE_REGION=us-east5
 # Blank to use the (default) database
 FIRESTORE_DATABASE=
+FIRESTORE_EMULATOR_HOST=localhost:8243
 
 # OpenTelemetry tracing configuration
 TRACE_AGENT_ENABLED=false
